+++ packages/analytics/analytics-chart/src/components/ZoomTimerange.vue
@@ -1,13 +1,19 @@
 <template>
   <div class="zoom-timerange-container">
     <div class="zoom-timerange-details">
-      <div class="label">
+      <div
+        class="label"
+        data-testid="zoom-timerange-from"
+      >
         {{ i18n.t('zoom_time_range.from') }}
       </div>
       <div>
         {{ throttledStartTime }}
       </div>
-      <div class="label">
+      <div
+        class="label"
+        data-testid="zoom-timerange-to"
+      >
         {{ i18n.t('zoom_time_range.to') }}
       </div>
       <div>
@@ -17,24 +23,37 @@
   </div>
 </template>
 <script setup lang="ts">
-import { formatTime } from '@kong-ui-public/analytics-utilities'
+import type { GranularityValues } from '@kong-ui-public/analytics-utilities'
 import composables from '../composables'
 import { ref, watch } from 'vue'
-import { debounce } from '../utils'
+import { debounce, formatTooltipTimestampByGranularity } from '../utils'
 
 const props = defineProps<{
   start: Date
   end: Date
+  granularity: GranularityValues
 }>()
 
 const { i18n } = composables.useI18n()
-const throttledStartTime = ref(formatTime(props.start.getTime()))
-const throttledEndTime = ref(formatTime(props.end.getTime()))
+const throttledStartTime = ref(formatTooltipTimestampByGranularity({
+  tickValue: props.start,
+  granularity: props.granularity,
+}))
+const throttledEndTime = ref(formatTooltipTimestampByGranularity({
+  tickValue: props.end,
+  granularity: props.granularity,
+}))
 
 const updateTimeRangeDebounced = (start: Date, end: Date) => {
   debounce(() => {
-    throttledStartTime.value = formatTime(start.getTime())
-    throttledEndTime.value = formatTime(end.getTime())
+    throttledStartTime.value = formatTooltipTimestampByGranularity({
+      tickValue: start,
+      granularity: props.granularity,
+    })
+    throttledEndTime.value = formatTooltipTimestampByGranularity({
+      tickValue: end,
+      granularity: props.granularity,
+    })
   }, 100)()
 }
 
@@ -46,6 +65,7 @@ watch(() => [props.start, props.end], ([newStart, newEnd]) => {
 
 <style scoped lang="scss">
 .zoom-timerange-container {
+  line-height: var(--kui-line-height-20, $kui-line-height-20);
   margin: 0 var(--kui-space-20, $kui-space-20);
 
   .zoom-timerange-details {
@@ -53,7 +73,7 @@ watch(() => [props.start, props.end], ([newStart, newEnd]) => {
     font-size: var(--kui-font-size-20, $kui-font-size-20);
     gap: var(--kui-space-10, $kui-space-10);
     grid-template-columns: 35px 1fr;
-    margin: var(--kui-space-30, $kui-space-30);
+    margin: var(--kui-space-30, $kui-space-30) var(--kui-space-10, $kui-space-10);
 
     .label {
       color: var(--kui-color-text-netural-strong, $kui-color-text-neutral-strong);