+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/NodeFormCall.vue
@@ -24,7 +24,7 @@
 
     <HttpMethodField
       name="method"
-      @update:model-value="setConfig()"
+      @update="setConfig()"
     />
 
     <NumberField