+++ packages/entities/entities-plugins/src/components/free-form/Datakit/flow-editor/node-forms/NodeFormProperty.vue
@@ -12,7 +12,7 @@
 
     <EnumField
       name="content_type"
-      @update:model-value="setConfig"
+      @update="setConfig()"
     />
 
     <PropertiesField