+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.2...@kong-ui-public/analytics-metric-provider@10.1.3) (2025-09-12)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies ([#2482](https://github.com/Kong/public-ui-components/issues/2482)) ([b45c9dc](https://github.com/Kong/public-ui-components/commit/b45c9dc6a6bbca72f72fa42829caad9bfb92402b))
+
+
+
+
+
+## [10.1.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.1...@kong-ui-public/analytics-metric-provider@10.1.2) (2025-09-12)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
+## [10.1.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.0...@kong-ui-public/analytics-metric-provider@10.1.1) (2025-09-12)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 # [10.1.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.0.2...@kong-ui-public/analytics-metric-provider@10.1.0) (2025-09-08)