+++ packages/entities/entities-upstreams-targets/CHANGELOG.md
@@ -3,6 +3,55 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.12.5](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.4...@kong-ui-public/entities-upstreams-targets@3.12.5) (2025-09-12)
+
+
+### Bug Fixes
+
+* **deps:** update all [@kong](https://github.com/kong) scoped dependencies ([#2482](https://github.com/Kong/public-ui-components/issues/2482)) ([b45c9dc](https://github.com/Kong/public-ui-components/commit/b45c9dc6a6bbca72f72fa42829caad9bfb92402b))
+
+
+
+
+
+## [3.12.4](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.3...@kong-ui-public/entities-upstreams-targets@3.12.4) (2025-09-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency axios to v1.12.0 [security] ([#2481](https://github.com/Kong/public-ui-components/issues/2481)) ([af05c92](https://github.com/Kong/public-ui-components/commit/af05c92131f666eb7ab89c27149d14d794b9535d))
+
+
+
+
+
+## [3.12.3](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.2...@kong-ui-public/entities-upstreams-targets@3.12.3) (2025-09-12)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
+## [3.12.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.1...@kong-ui-public/entities-upstreams-targets@3.12.2) (2025-09-10)
+
+
+### Bug Fixes
+
+* **entities-upstreams-targets:** edit state issue ([#2470](https://github.com/Kong/public-ui-components/issues/2470)) ([9bc8de9](https://github.com/Kong/public-ui-components/commit/9bc8de9cf459960070428c942b7cf7aae4707159))
+
+
+
+
+
+## [3.12.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.12.0...@kong-ui-public/entities-upstreams-targets@3.12.1) (2025-09-10)
+
+**Note:** Version bump only for package @kong-ui-public/entities-upstreams-targets
+
+
+
+
+
 # [3.12.0](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/entities-upstreams-targets@3.11.4...@kong-ui-public/entities-upstreams-targets@3.12.0) (2025-09-08)